bur_form/requirements.txt
asgiref==3.7.2
Django==4.2.2
gunicorn==20.1.0
django-environ==0.11.2
django-mptt==0.14.0
django-cleanup==8.0.0
django-ckeditor-5==0.2.9
django-taggit==4.0.0
django-recaptcha==3.0.0
django_debug_toolbar==4.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.6.3
Pillow==10.0.1
pytils==0.4.1
django-recaptcha==3.0.0